requirements.txt
pdfplumber==0.11.0
python-docx==1.1.0
openai>=1.50.0
httpx<0.28.0
pandas==2.2.1
openpyxl==3.1.2
pydantic==2.6.3
tqdm==4.66.2
